+++ board/bootstub_declarations.c
@@ -1,0 +1,5 @@
+#include "bootstub_declarations.h"
+
+// ********************* Globals **********************
+uint8_t hw_type = 0;
+board *current_board;
