+++ board/comms_definitions.h
@@ -1,3 +1,7 @@
+#pragma once
+
+#include <stdint.h>
+
 typedef struct {
   uint8_t request;
   uint16_t param1;
