+++ board/faults.h
@@ -1,3 +1,7 @@
+#pragma once
+
+#include <stdint.h>
+
 #define FAULT_STATUS_NONE 0U
 #define FAULT_STATUS_TEMPORARY 1U
 #define FAULT_STATUS_PERMANENT 2U
@@ -34,26 +38,8 @@
 // Permanent faults
 #define PERMANENT_FAULTS 0U
 
-uint8_t fault_status = FAULT_STATUS_NONE;
-uint32_t faults = 0U;
+extern uint8_t fault_status;
+extern uint32_t faults;
 
-void fault_occurred(uint32_t fault) {
-  if ((faults & fault) == 0U) {
-    if ((PERMANENT_FAULTS & fault) != 0U) {
-      print("Permanent fault occurred: 0x"); puth(fault); print("\n");
-      fault_status = FAULT_STATUS_PERMANENT;
-    } else {
-      print("Temporary fault occurred: 0x"); puth(fault); print("\n");
-      fault_status = FAULT_STATUS_TEMPORARY;
-    }
-  }
-  faults |= fault;
-}
-
-void fault_recovered(uint32_t fault) {
-  if ((PERMANENT_FAULTS & fault) == 0U) {
-    faults &= ~fault;
-  } else {
-    print("Cannot recover from a permanent fault!\n");
-  }
-}
+void fault_occurred(uint32_t fault);
+void fault_recovered(uint32_t fault);
