+++ tests/libpanda/panda.c
@@ -1,12 +1,12 @@
 #include "fake_stm.h"
 #include "config.h"
 #include "can_definitions.h"
+#include "drivers/harness.h"
 
 bool can_init(uint8_t can_number) { return true; }
 void process_can(uint8_t can_number) { }
 //int safety_tx_hook(CANPacket_t *to_send) { return 1; }
 
-typedef struct harness_configuration harness_configuration;
 void refresh_can_tx_slots_available(void);
 void can_tx_comms_resume_usb(void) { };
 void can_tx_comms_resume_spi(void) { };
